bot/requirements.txt
uvicorn
fastapi
aiogram
pydantic
asyncio
python-dotenv